pyproject.toml
[project]
name = "demo-erc8004"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "José Cabrero-Holgueras", email = "jose.cabrero@nillion.com" }
]
requires-python = ">=3.12"
dependencies = [
    "web3>=7.0.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["uv_build>=0.9.4,<0.10.0"]
build-backend = "uv_build"
